separated tests of sym_wann into one file, added a bit more tests
