+++ tests/common_systems.py
@@ -161,6 +161,7 @@ def system_GaAs_W90(create_files_GaAs_W90):
     # Load system
     seedname = os.path.join(data_dir, "GaAs")
     system = wberri.system.System_w90(seedname, berry=True, morb=True, spin=True, transl_inv=False)
+    system.set_symmetry(symmetries_GaAs)
 
     return system
 
@@ -174,6 +175,7 @@ def system_GaAs_W90_wcc(create_files_GaAs_W90):
     # Load system
     seedname = os.path.join(data_dir, "GaAs")
     system = wberri.system.System_w90(seedname, morb=True, transl_inv=False, spin=True, use_wcc_phase=True)
+    system.set_symmetry(symmetries_GaAs)
 
     return system
 
@@ -190,6 +192,7 @@ def system_GaAs_tb():
 
     seedname = os.path.join(data_dir, "GaAs_tb.dat")
     system = wberri.system.System_tb(seedname, berry=True)
+    system.set_symmetry(symmetries_GaAs)
 
     return system
 
@@ -228,7 +231,7 @@ def system_GaAs_tb_wcc():
     # Load system
     seedname = os.path.join(data_dir, "GaAs_tb.dat")
     system = wberri.system.System_tb(seedname, berry=True, use_wcc_phase=True)
-
+    system.set_symmetry(symmetries_GaAs)
     return system
