+++ tests/test_run.py
@@ -384,62 +384,6 @@ def test_Fe_sym(check_run, system_Fe_W90, compare_any_result):
             result_type=EnergyResult)
 
 
-def test_Fe_sym_W90(check_run, system_Fe_sym_W90, compare_any_result):
-    param = {'Efermi': Efermi_Fe}
-    cals = {'ahc': calc.static.AHC,
-            'Morb': calc.static.Morb,
-            'spin': calc.static.Spin}
-    calculators = {k: v(**param) for k, v in cals.items()}
-    check_run(
-        system_Fe_sym_W90,
-        calculators,
-        fout_name="berry_Fe_sym_W90",
-        suffix="-run",
-        use_symmetry=False
-    )
-    cals = {'gyrotropic_Korb': calc.static.GME_orb_FermiSea,
-            'berry_dipole': calc.static.BerryDipole_FermiSea,
-            'gyrotropic_Kspin': calc.static.GME_spin_FermiSea}
-    calculators = {k: v(**param) for k, v in cals.items()}
-    check_run(
-        system_Fe_sym_W90,
-        calculators,
-        fout_name="berry_Fe_sym_W90",
-        precision=1e-8,
-        suffix="-run",
-        compare_zero=True,
-        use_symmetry=False
-    )
-
-
-def test_Fe_sym_W90_sym(check_run, system_Fe_sym_W90, compare_any_result):
-    param = {'Efermi': Efermi_Fe}
-    cals = {'ahc': calc.static.AHC,
-            'Morb': calc.static.Morb,
-            'spin': calc.static.Spin}
-    calculators = {k: v(**param) for k, v in cals.items()}
-    check_run(
-        system_Fe_sym_W90,
-        calculators,
-        fout_name="berry_Fe_sym_W90",
-        suffix="sym-run",
-        use_symmetry=True
-    )
-    cals = {'gyrotropic_Korb': calc.static.GME_orb_FermiSea,
-            'berry_dipole': calc.static.BerryDipole_FermiSea,
-            'gyrotropic_Kspin': calc.static.GME_spin_FermiSea}
-    calculators = {k: v(**param) for k, v in cals.items()}
-    check_run(
-        system_Fe_sym_W90,
-        calculators,
-        fout_name="berry_Fe_sym_W90",
-        suffix="sym-run",
-        precision=1e-8,
-        compare_zero=True,
-        use_symmetry=True
-    )
-
-
 def test_Fe_FPLO(check_run, system_Fe_FPLO, compare_any_result):
     param = {'Efermi': Efermi_Fe_FPLO}
     calculators = {k: v(**param) for k, v in calculators_Fe.items()}
@@ -752,17 +696,6 @@ def test_GaAs_tb_wcc_ws(check_run, system_GaAs_tb_wcc_ws, compare_any_result):
     )  # This is a low precision for the nonabelian thing, not sure if it does not indicate a problem, or is a gauge-dependent thing
 
 
-def test_GaAs_sym_tb(check_run, system_GaAs_sym_tb, compare_any_result):
-
-    check_run(
-        system_GaAs_sym_tb,
-        {'ahc': calc.static.AHC(Efermi=Efermi_GaAs)},
-        fout_name="berry_GaAs_sym_tb",
-        precision=1e-5,
-        compare_zero=True,
-        suffix="run",
-    )
-
 
 def test_Haldane_PythTB(check_run, system_Haldane_PythTB, compare_any_result):
 
@@ -807,57 +740,6 @@ def test_GaAs_dynamic(check_run, system_GaAs_W90, compare_any_result):
     )
 
 
-def test_GaAs_dynamic_sym(check_run, system_GaAs_sym_tb, compare_any_result):
-    "Test shift current and injection current"
-
-    param = dict(
-        Efermi=Efermi_GaAs,
-        omega=np.arange(1.0, 5.1, 0.5),
-        smr_fixed_width=0.2,
-        smr_type='Gaussian',
-        kBT=0.01,
-    )
-    calculators = dict(
-        shift_current=calc.dynamic.ShiftCurrent(sc_eta=0.1, **param),
-        injection_current=calc.dynamic.InjectionCurrent(**param),
-        opt_conductivity=wberri.calculators.dynamic.OpticalConductivity(**param)
-    )
-
-    result_full_k = check_run(
-        system_GaAs_sym_tb,
-        calculators,
-        fout_name="dynamic_GaAs_sym",
-        grid_param={
-            'NK': [6, 6, 6],
-            'NKFFT': [3, 3, 3]
-        },
-        use_symmetry=False,
-        do_not_compare=True,
-            )
-
-    result_irr_k = check_run(
-        system_GaAs_sym_tb,
-        calculators,
-        fout_name="dynamic_GaAs_sym",
-        suffix="sym",
-        suffix_ref="",
-        grid_param={
-            'NK': [6, 6, 6],
-            'NKFFT': [3, 3, 3]
-        },
-        use_symmetry=True,
-        do_not_compare=True,
-    )
-
-
-    assert result_full_k.results["shift_current"].data == approx(
-        result_irr_k.results["shift_current"].data, abs=1e-6)
-
-    assert result_full_k.results["injection_current"].data == approx(
-        result_irr_k.results["injection_current"].data, abs=1e-6)
-
-    assert result_full_k.results["opt_conductivity"].data == approx(
-        result_irr_k.results["opt_conductivity"].data, abs=1e-7)
 
 def test_Haldane_TBmodels(check_run, system_Haldane_TBmodels, compare_any_result):
